Add functionality to delete children
Add ChildContainerKey type to HasChildren, which allows for a uniform
interface to select a child to delete between layers.

Add deleteChild function to HasChildren, which runs a cleanup function
and deletes the specified key from the ChildContainer. This is meant to
stop any Watch and Event instances before deleting the child.

Signed-off-by: hololeap <hololeap@protonmail.com>